Update DMA and ADC settings in IOC file
DMA transfer from memory to SPI1 needed to be set to halfword data size (two bytes) to accommodate for SPI changing to 16 bit length when display framebuffer flush occurs. Also, ADC clock prescaler was set to /4 to yield an ADC clock of 160/4=40 MHz, because the ADC has a max clock frequency of 55 MHz according to the datasheet! (Not sure how it let me use the ADC with the clock set to 160 MHz... lol.